Reorganize code into separate files
