+++ cli/utils.go
@@ -1,6 +1,7 @@
 package cli
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -249,7 +250,44 @@ func randDur(window time.Duration) time.Duration {
 	return window + time.Duration(rand.Int63n(int64(window)))
 }
 
-// validCommand checks that a command is a recognized CLI command.
-func validateCLIArgs(args []string) {
-	// Ignore first args
+// CappedWriter implements a buffer that stores up to MaxBytes bytes.
+// Returns ErrBufferFull on overflowing writes.
+type CappedWriter struct {
+	Buffer   *bytes.Buffer
+	MaxBytes int
 }
+
+// ErrBufferFull returns an error indicating that a CappedWriter's buffer has
+// reached max capacity.
+func ErrBufferFull() error {
+	return errors.New("buffer is full")
+}
+
+// Write writes to the buffer. An error is returned if the buffer is full.
+func (cw *CappedWriter) Write(in []byte) (int, error) {
+	remaining := cw.MaxBytes - cw.Buffer.Len()
+	if len(in) <= remaining {
+		return cw.Buffer.Write(in)
+	}
+	// fill up the remaining buffer and return an error
+	n, _ := cw.Buffer.Write(in[:remaining])
+	return n, ErrBufferFull()
+}
+
+// IsFull indicates whether the buffer is full.
+func (cw *CappedWriter) IsFull() bool {
+	return cw.Buffer.Len() == cw.MaxBytes
+}
+
+// String return the contents of the buffer as a string.
+func (cw *CappedWriter) String() string {
+	return cw.Buffer.String()
+}
+
+func (cw *CappedWriter) Bytes() []byte {
+	return cw.Buffer.Bytes()
+}
+
+// Close is a noop method so that you can use CappedWriter as an
+// io.WriteCloser.
+func (cw *CappedWriter) Close() error { return nil }
